+++ pytest_drf/util/__init__.py
@@ -1,2 +1,3 @@
 from .expressions import *
 from .metaclasses import *
+from .urls import *